requirements.txt
transformers>=4.40.0
requests>=2.31.0
python-dotenv>=1.0.0
uvicorn== 0.37.0
fastapi==0.118.0
pydantic==2.11.10
torch==2.8.0
transformers==4.57.0
accelerate
huggingface_hub[hf_xet]
scikit-learn==1.2.2
numpy==1.24.3
scipy==1.10.1
prometheus-fastapi-instrumentator
joblib

